going to just pretty much give up on this one and move on
getiting a syntax error static_cast when trying to pass a pointer of type
QApplication to the CustomWidget constructor.  I was doing this so that I
could call the QApplication version of quit() when the button instide the
custom widget is clicked but apparently everything just wants to have
aids.